type convertion

+++ Sparky/src/graphics/postfx/PostEffectsPass.cpp
@@ -19,7 +19,7 @@ namespace sparky
 		void PostEffectsPass::RenderPass(Framebuffer *target)
 		{
 			m_Shader->Bind();
-			m_Shader->SetUniform("pr_matrix", maths::mat4::Orthographic(0, target->GetWidth(), target->GetHeight(), 0, -1.0f, 1.0f));
+			m_Shader->SetUniform("pr_matrix", maths::mat4::Orthographic(0.0f, (float)target->GetWidth(), (float)target->GetHeight(), 0.0f, -1.0f, 1.0f));
 			GLCall(glDrawElements(GL_TRIANGLES, 6, GL_UNSIGNED_INT, NULL));
 			m_Shader->Unbind();
 		}
